+++ tcp/proxy/proxy.go
@@ -1,51 +1,27 @@
-package main
+package tcp
 
 import (
+	"fmt"
 	"io"
 	"log"
 	"net"
 )
 
-// ProxyServer creates a TCP base Proxy
-func ProxyServer(conn net.Conn) {
-
-	defer conn.Close()
-	b := make([]byte, 512)
-	for {
-		size, err := conn.Read(b[0:])
-		if err == io.EOF {
-			log.Println("Client disconnected.")
-			break
-		}
-		if err != nil {
-			log.Println("Unexpected error.")
-			break
-		}
-
-		log.Printf("Received %d bytes: %s\n", size, string(b))
-
-		// now send the data
-		log.Println("Writing data")
-		if _, err := conn.Write(b[0:size]); err != nil {
-			log.Fatalln("Unable to write data.")
-		}
-
-	}
-
-}
-
-func main() {
-	listener, err := net.Listen("tcp", ":20080")
+// ProxyServer creates a TCP based proxy server
+func ProxyServer(src net.Conn) {
+	dst, err := net.Dial("tcp", "localhost:3001")
 	if err != nil {
-		log.Fatalln("Unable to bind to port.")
+		fmt.Println("Unable to connect to our unreachable host.")
 	}
-	log.Println("Listening on 0.0.0.0:20080")
-	for {
-		conn, err := listener.Accept()
-		log.Println("Received connection")
-		if err != nil {
-			log.Fatalln("Unable to accept connection")
+	defer dst.Close()
+	go func() {
+		if _, err := io.Copy(dst, src); err != nil {
+			log.Fatalln(err)
 		}
-		go ProxyServer(conn)
+	}()
+
+	if _, err := io.Copy(src, dst); err != nil {
+		log.Fatalln(err)
 	}
+
 }
