--- tcp/proxy/proxy_test.go
@@ -1,9 +1,0 @@
-package tcp
-
-import (
-	"testing"
-)
-
-func TestProxyServer(t *testing.T) {
-
-}
